second commit

+++ src/main/java/com/app/sony/entity/ProductEntity.java
@@ -1,5 +1,6 @@
 package com.app.sony.entity;
 
+import io.swagger.models.auth.In;
 import jakarta.persistence.*;
 
 import java.util.Objects;
@@ -11,7 +12,7 @@ public class ProductEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name", nullable = false)
     private String name;
@@ -35,11 +36,11 @@ public class ProductEntity {
     @Column(name = "image_url", nullable = false)
     private String imageUrl;
 
-    public Long getId() {
+    public Integer getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(Integer id) {
         this.id = id;
     }
